Update
Added testing files and functional changes

+++ Schools/School_C/Models/Score.cs
@@ -5,7 +5,7 @@ namespace School.Models
         public int ScoreID { get; set; }
         public int RollNumber { get; set; }
         public int SubjectID { get; set; }
-        public decimal Mark { get; set; }
+        public float Mark { get; set; }
         public Student? student { get; set; }
         public Subject? subject { get; set; }
